builtins: move to own submodule, fixes #1177
