+++ devito/builtins/__init__.py
@@ -1,0 +1,2 @@
+from .arithmetic import *  # noqa
+from .initializers import *  # noqa
