+++ devito/builtins/initializers.py
@@ -1,15 +1,10 @@
-"""
-Built-in Operators provided by Devito.
-"""
-
-from sympy import Abs, Pow
 import numpy as np
 
 import devito as dv
 from devito.tools import as_tuple, as_list
+from devito.builtins.utils import nbl_to_padsize, pad_outhalo
 
-__all__ = ['assign', 'smooth', 'gaussian_smooth', 'initialize_function', 'norm',
-           'sumall', 'inner', 'mmin', 'mmax']
+__all__ = ['assign', 'smooth', 'gaussian_smooth', 'initialize_function']
 
 
 def assign(f, rhs=0, options=None, name='assign', **kwargs):
@@ -205,31 +200,8 @@ def gaussian_smooth(f, sigma=1, truncate=4.0, mode='reflect'):
     return f
 
 
-def nbl_to_padsize(nbl, ndim):
-    """
-    Creates the pad sizes from `nbl`. The output is a tuple of tuple
-    such as ((40, 40), (40, 40)) where each subtuple is the left/right
-    pad size for the dimension.
-    """
-    nb_total = as_tuple(nbl)
-    if len(nb_total) == 1 and len(nb_total) < ndim:
-        nb_pad = ndim*((nb_total[0], nb_total[0]),)
-    elif len(nb_total) == ndim:
-        nb_pad = []
-        for nb_dim in nb_total:
-            if len(as_tuple(nb_dim)) == 1:
-                nb_pad.append((nb_dim, nb_dim))
-            else:
-                assert len(nb_dim) == 2
-                nb_pad.append(nb_dim)
-    else:
-        raise ValueError("`nbl` must be an integer or tuple of (tuple of) integers"
-                         "of length `function.ndim`.")
-    return tuple(nb_pad)
-
-
 def initialize_function(function, data, nbl, mapper=None, mode='constant',
-                        name='padfunc', pad_halo=True, **kwargs):
+                        name=None, pad_halo=True, **kwargs):
     """
     Initialize a Function with the given ``data``. ``data``
     does *not* include the ``nbl`` outer/boundary layers; these are added via padding
@@ -256,7 +228,7 @@ def initialize_function(function, data, nbl, mapper=None, mode='constant',
     name : str, optional
         The name assigned to the operator.
     pad_halo : bool, optional
-        Whether to also pad the finite-difference halo.
+        Whether to also pad the outer halo.
 
     Examples
     --------
@@ -298,6 +270,7 @@ def initialize_function(function, data, nbl, mapper=None, mode='constant',
           [2, 3, 3, 3, 3, 2],
           [2, 2, 2, 2, 2, 2]], dtype=int32)
     """
+    name = name or 'pad_%s' % function.name
     if isinstance(function, dv.TimeFunction):
         raise NotImplementedError("TimeFunctions are not currently supported.")
 
@@ -307,7 +280,7 @@ def initialize_function(function, data, nbl, mapper=None, mode='constant',
         else:
             function.data[:] = data[:]
         if pad_halo:
-            pad_fd_halo(function)
+            pad_outhalo(function)
         return
 
     nbl = nbl_to_padsize(nbl, function.ndim)
@@ -368,218 +341,4 @@ def initialize_function(function, data, nbl, mapper=None, mode='constant',
     assign(lhs, rhs, options=options, name=name, **kwargs)
 
     if pad_halo:
-        pad_fd_halo(function)
-
-
-def pad_fd_halo(function):
-    """ Pad finite-difference halo with edge values."""
-    h_shape = function._data_with_outhalo.shape
-    for i, h in enumerate(function._size_outhalo):
-        slices = [slice(None)]*len(function.shape)
-        slices_d = [slice(None)]*len(function.shape)
-        if h.left != 0:
-            # left part
-            slices[i] = slice(0, h.left)
-            slices_d[i] = slice(h.left, h.left+1, 1)
-            function._data_with_outhalo._local[tuple(slices)] \
-                = function._data_with_outhalo._local[tuple(slices_d)]
-        if h.right != 0:
-            # right part
-            slices[i] = slice(h_shape[i] - h.right, h_shape[i], 1)
-            slices_d[i] = slice(h_shape[i] - h.right - 1, h_shape[i] - h.right, 1)
-            function._data_with_outhalo._local[tuple(slices)] \
-                = function._data_with_outhalo._local[tuple(slices_d)]
-
-
-# Reduction-inducing builtins
-
-class MPIReduction(object):
-    """
-    A context manager to build MPI-aware reduction Operators.
-    """
-
-    def __init__(self, *functions, op=dv.mpi.MPI.SUM):
-        grids = {f.grid for f in functions}
-        if len(grids) == 0:
-            self.grid = None
-        elif len(grids) == 1:
-            self.grid = grids.pop()
-        else:
-            raise ValueError("Multiple Grids found")
-        dtype = {f.dtype for f in functions}
-        if len(dtype) == 1:
-            self.dtype = dtype.pop()
-        else:
-            raise ValueError("Illegal mixed data types")
-        self.v = None
-        self.op = op
-
-    def __enter__(self):
-        i = dv.Dimension(name='i',)
-        self.n = dv.Function(name='n', shape=(1,), dimensions=(i,),
-                             grid=self.grid, dtype=self.dtype)
-        self.n.data[0] = 0
-        return self
-
-    def __exit__(self, exc_type, exc_value, traceback):
-        if self.grid is None or not dv.configuration['mpi']:
-            assert self.n.data.size == 1
-            self.v = self.n.data[0]
-        else:
-            comm = self.grid.distributor.comm
-            self.v = comm.allreduce(np.asarray(self.n.data), self.op)[0]
-
-
-@dv.switchconfig(log_level='ERROR')
-def norm(f, order=2):
-    """
-    Compute the norm of a Function.
-
-    Parameters
-    ----------
-    f : Function
-        Input Function.
-    order : int, optional
-        The order of the norm. Defaults to 2.
-    """
-    kwargs = {}
-    if f.is_TimeFunction and f._time_buffering:
-        kwargs[f.time_dim.max_name] = f._time_size - 1
-
-    # Protect SparseFunctions from accessing duplicated (out-of-domain) data,
-    # otherwise we would eventually be summing more than expected
-    p, eqns = f.guard() if f.is_SparseFunction else (f, [])
-
-    s = dv.types.Scalar(name='sum', dtype=f.dtype)
-
-    with MPIReduction(f) as mr:
-        op = dv.Operator([dv.Eq(s, 0.0)] +
-                         eqns +
-                         [dv.Inc(s, Abs(Pow(p, order))), dv.Eq(mr.n[0], s)],
-                         name='norm%d' % order)
-        op.apply(**kwargs)
-
-    v = Pow(mr.v, 1/order)
-
-    return f.dtype(v)
-
-
-def sumall(f):
-    """
-    Compute the sum of all Function data.
-
-    Parameters
-    ----------
-    f : Function
-        Input Function.
-    """
-    kwargs = {}
-    if f.is_TimeFunction and f._time_buffering:
-        kwargs[f.time_dim.max_name] = f._time_size - 1
-
-    # Protect SparseFunctions from accessing duplicated (out-of-domain) data,
-    # otherwise we would eventually be summing more than expected
-    p, eqns = f.guard() if f.is_SparseFunction else (f, [])
-
-    s = dv.types.Scalar(name='sum', dtype=f.dtype)
-
-    with MPIReduction(f) as mr:
-        op = dv.Operator([dv.Eq(s, 0.0)] +
-                         eqns +
-                         [dv.Inc(s, p), dv.Eq(mr.n[0], s)],
-                         name='sum')
-        op.apply(**kwargs)
-
-    return f.dtype(mr.v)
-
-
-def inner(f, g):
-    """
-    Inner product of two Functions.
-
-    Parameters
-    ----------
-    f : Function
-        First input operand
-    g : Function
-        Second input operand
-
-    Raises
-    ------
-    ValueError
-        If the two input Functions are defined over different grids, or have
-        different dimensionality, or their dimension-wise sizes don't match.
-        If in input are two SparseFunctions and their coordinates don't match,
-        the exception is raised.
-
-    Notes
-    -----
-    The inner product is the sum of all dimension-wise products. For 1D Functions,
-    the inner product corresponds to the dot product.
-    """
-    # Input check
-    if f.is_TimeFunction and f._time_buffering != g._time_buffering:
-        raise ValueError("Cannot compute `inner` between save/nosave TimeFunctions")
-    if f.shape != g.shape:
-        raise ValueError("`f` and `g` must have same shape")
-    if f._data is None or g._data is None:
-        raise ValueError("Uninitialized input")
-    if f.is_SparseFunction and not np.all(f.coordinates_data == g.coordinates_data):
-        raise ValueError("Non-matching coordinates")
-
-    kwargs = {}
-    if f.is_TimeFunction and f._time_buffering:
-        kwargs[f.time_dim.max_name] = f._time_size - 1
-
-    # Protect SparseFunctions from accessing duplicated (out-of-domain) data,
-    # otherwise we would eventually be summing more than expected
-    rhs, eqns = f.guard(f*g) if f.is_SparseFunction else (f*g, [])
-
-    s = dv.types.Scalar(name='sum', dtype=f.dtype)
-
-    with MPIReduction(f, g) as mr:
-        op = dv.Operator([dv.Eq(s, 0.0)] +
-                         eqns +
-                         [dv.Inc(s, rhs), dv.Eq(mr.n[0], s)],
-                         name='inner')
-        op.apply(**kwargs)
-
-    return f.dtype(mr.v)
-
-
-def mmin(f):
-    """
-    Retrieve the minimum.
-
-    Parameters
-    ----------
-    f : array_like or Function
-        Input operand.
-    """
-    if isinstance(f, dv.Constant):
-        return f.data
-    elif isinstance(f, dv.types.dense.DiscreteFunction):
-        with MPIReduction(f, op=dv.mpi.MPI.MIN) as mr:
-            mr.n.data[0] = np.min(f.data_ro_domain).item()
-        return mr.v.item()
-    else:
-        raise ValueError("Expected Function, not `%s`" % type(f))
-
-
-def mmax(f):
-    """
-    Retrieve the maximum.
-
-    Parameters
-    ----------
-    f : array_like or Function
-        Input operand.
-    """
-    if isinstance(f, dv.Constant):
-        return f.data
-    elif isinstance(f, dv.types.dense.DiscreteFunction):
-        with MPIReduction(f, op=dv.mpi.MPI.MAX) as mr:
-            mr.n.data[0] = np.max(f.data_ro_domain).item()
-        return mr.v.item()
-    else:
-        raise ValueError("Expected Function, not `%s`" % type(f))
+        pad_outhalo(function)
